+++ client/src/hooks/useLanguage.tsx
@@ -3,6 +3,7 @@ import moment from 'moment';
 import 'moment/locale/de';
 import 'moment/locale/es';
 import 'moment/locale/fr';
+import 'moment/locale/ru';
 import React, { useContext, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAlert } from './useAlert';