+++ client/src/utils/Server.ts
@@ -1,4 +1,4 @@
-import { BrutesCreateResponse, BrutesExistsResponse, BrutesGetDestinyResponse, BrutesGetFightsLeftResponse, BrutesGetForRankResponse, BrutesGetOpponentsResponse, BrutesGetRankingResponse, BruteWithBodyColors, FullTournament, ServerReadyResponse, TournamentsGetGlobalResponse, UserWithBrutesBodyColor } from '@labrute/core';
+import { BrutesCreateResponse, BrutesExistsResponse, BrutesGetDestinyResponse, BrutesGetFightsLeftResponse, BrutesGetForRankResponse, BrutesGetOpponentsResponse, BrutesGetRankingResponse, BruteWithBodyColors, FullTournament, ServerReadyResponse, TournamentsGetGlobalResponse, UsersAdminUpdateRequest, UserWithBrutesBodyColor } from '@labrute/core';
 import { Achievement, Brute, DestinyChoice, DestinyChoiceSide, Fight, Gender, Lang, Log, Prisma, Tournament, User } from '@labrute/prisma';
 import Fetch from './Fetch';
 
@@ -9,11 +9,16 @@ const Server = {
       login,
       token
     }, 'POST'),
-    list: () => Fetch<User[]>('/api/user/list', {}, 'GET'),
+    get: ({
+      name,
+      include,
+      where,
+    }: { name: string, include?: Prisma.UserInclude, where?: Prisma.UserWhereInput }) => Fetch<User>(`/api/user/${name}/get`, { include, where }, 'POST'),
     runDailyJob: () => Fetch<never>('/api/run-daily-job'),
     changeLanguage: (lang: Lang) => Fetch<never>('/api/user/change-language', { lang }, 'POST'),
     changeFightSpeed: (fightSpeed: number) => Fetch<never>('/api/user/change-fight-speed', { fightSpeed }, 'POST'),
     toggleBackgroundMusic: (backgroundMusic: boolean) => Fetch<never>('/api/user/toggle-background-music', { backgroundMusic }, 'POST'),
+    adminUpdate: (id: string, data: UsersAdminUpdateRequest) => Fetch<never>(`/api/user/${id}/admin-update`, data, 'POST'),
   },
   Brute: {
     list: () => Fetch<Brute[]>('/api/brute/list/'),
@@ -46,7 +51,7 @@ const Server = {
     levelUp: (
       name: string,
       choice: DestinyChoiceSide,
-    ) => Fetch<never>(`/api/brute/${name}/level-up`, { choice }, 'POST'),
+    ) => Fetch<Brute>(`/api/brute/${name}/level-up`, { choice }, 'POST'),
     getOpponents: (name: string, level: number) => Fetch<BrutesGetOpponentsResponse>(`/api/brute/${name}/get-opponents/${level}`),
     sacrifice: (name: string) => Fetch<{ points: number }>(`/api/brute/${name}/sacrifice`, {}, 'GET'),
     getForRank: ({ name, rank }: { name: string, rank?: number }) => Fetch<BrutesGetForRankResponse>(`/api/brute/${name}/ranking-data${typeof rank === 'undefined' ? '' : `/${rank}`}`),