+++ server/src/routes.ts
@@ -31,10 +31,12 @@ const initRoutes = (app: Express, prisma: PrismaClient) => {
 
   // User
   app.post('/api/user/authenticate', Users.authenticate(prisma));
+  app.post('/api/user/:name/get', Users.get(prisma));
   app.get('/api/run-daily-job', Users.runDailyJob(prisma));
   app.post('/api/user/change-language', Users.changeLanguage(prisma));
   app.post('/api/user/change-fight-speed', Users.changeFightSpeed(prisma));
   app.post('/api/user/toggle-background-music', Users.toggleBackgroundMusic(prisma));
+  app.post('/api/user/:id/admin-update', Users.adminUpdate(prisma));
 
   // Brute
   app.get('/api/brute/list', Brutes.list(prisma));