+++ server/src/utils/fight/getDamage.ts
@@ -8,14 +8,19 @@ const getDamage = (
   const base = fighter.activeWeapon?.damage || fighter.baseDamage;
   let skillsMultiplier = 1;
 
+  // Using Piledriver ?
+  const piledriver = fighter.activeSkills.find((sk) => sk.name === 'hammer');
+
   // +50% damage for `weaponsMaster` on sharp weapons
   if (fighter.activeWeapon?.types.includes('sharp') && fighter.skills.find((sk) => sk.name === 'weaponsMaster')) {
     skillsMultiplier += 0.5;
   }
 
-  // +100% damage for `martialArts` without a weapon or with a mug
-  if ((!fighter.activeWeapon || fighter.activeWeapon.name === 'mug') && fighter.skills.find((sk) => sk.name === 'martialArts')) {
-    skillsMultiplier += 1;
+  if (!piledriver) {
+    // +100% damage for `martialArts` without a weapon or with a mug
+    if ((!fighter.activeWeapon || fighter.activeWeapon.name === 'mug') && fighter.skills.find((sk) => sk.name === 'martialArts')) {
+      skillsMultiplier += 1;
+    }
   }
 
   // -30% damage if opponent has `leadSkeleton` and weapon is blunt
@@ -28,8 +33,8 @@ const getDamage = (
     skillsMultiplier *= 2;
   }
 
-  // x4 damage for `hammer`
-  if (fighter.activeSkills.find((sk) => sk.name === 'hammer')) {
+  // x4 damage for `piledriver`
+  if (piledriver) {
     skillsMultiplier *= 4;
   }
 
@@ -40,7 +45,7 @@ const getDamage = (
       (base + fighter.strength * 0.1 + fighter.agility * 0.15)
       * (1 + Math.random() * 0.5),
     );
-  } else if (fighter.activeSkills.find((sk) => sk.name === 'hammer')) {
+  } else if (piledriver) {
     damage = Math.floor(
       (10 + fighter.strength * 0.6)
       * (0.8 + Math.random() * 0.4) * skillsMultiplier,
@@ -52,8 +57,10 @@ const getDamage = (
     );
   }
 
-  // Reduce damage with opponent's armor
-  damage -= opponent.armor;
+  // Reduce damage with opponent's armor if not thrown
+  if (!thrown) {
+    damage -= opponent.armor;
+  }
 
   // Set minimum damage to 1
   if (damage < 1) {